// src/Models/ProviderType.tsx
 export interface ProviderType{

 
 id: number;

  age: number;
  

  userName?: string;
  password?: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  dob?: string;
  phoneNumber?: string;
  village?: string;
  taluk?: string;
  district?: string;
  pincode?: number;

cowCount:number;
aadharNumber:string;
bankName:string;
bankAccountNumber:string;

//  milkCapacityPerDay?: number; 
//   isActive?: boolean;
 }